(PC-4203): Inline offers request in paginated_offers_sql_repository and remove dictify from domain
